+++ agenda-connectors-services/src/main/resources/conf/portal/configuration.xml
@@ -15,16 +15,16 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 -->
-<configuration xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd http://www.exoplatform.org/xml/ns/kernel_1_2.xsd" xmlns="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd">
+<configuration xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.exoplatform.org/xml/ns/kernel_1_3.xsd http://www.exoplatform.org/xml/ns/kernel_1_3.xsd" xmlns="http://www.exoplatform.org/xml/ns/kernel_1_3.xsd">
 
-    <component>
-        <key>org.exoplatform.agendaconnector.service.ExchangeConnectorService</key>
-        <type>org.exoplatform.agendaconnector.service.ExchangeConnectorServiceImpl</type>
-    </component>
-    <component>
-        <type>org.exoplatform.agendaconnector.rest.ExchangeREST</type>
-    </component>
-    <component>
-        <type>org.exoplatform.agendaconnector.storage.ExchangeConnectorStorage</type>
-    </component>
+  <component>
+    <key>org.exoplatform.agendaconnector.service.ExchangeConnectorService</key>
+    <type>org.exoplatform.agendaconnector.service.ExchangeConnectorServiceImpl</type>
+  </component>
+  <component>
+    <type>org.exoplatform.agendaconnector.rest.ExchangeConnectorRest</type>
+  </component>
+  <component>
+    <type>org.exoplatform.agendaconnector.storage.ExchangeConnectorStorage</type>
+  </component>
 </configuration>
